refactor(cucumber): extract shared formatter helpers to lib
